refactor : RecordService에서 FlaskService, HistoryService 분리
