+++ src/main/java/plate/back/domain/predictedPlate/entity/PredictedPlate.java
@@ -13,11 +13,14 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import plate.back.domain.record.entity.Record;
 import plate.back.global.flask.dto.ModelPredictResult;
 
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Entity
 public class PredictedPlate {
