+++ src/main/java/plate/back/domain/record/dto/MultiResponseDto.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import lombok.Builder;
 import lombok.Getter;
-import plate.back.domain.image.dto.ImageResponseDto;
 import plate.back.domain.predictedPlate.dto.PredictedPlateDto;
 
 @Getter
@@ -13,15 +12,12 @@ public class MultiResponseDto {
 
     private final RecordResponseDto record;
     private final List<PredictedPlateDto> predictedList;
-    private final ImageResponseDto plateImage;
 
-    public static MultiResponseDto of(RecordResponseDto record, List<PredictedPlateDto> predictedList,
-            ImageResponseDto plateImage) {
+    public static MultiResponseDto of(RecordResponseDto record, List<PredictedPlateDto> predictedList) {
 
         return MultiResponseDto.builder()
                 .record(record)
                 .predictedList(predictedList)
-                .plateImage(plateImage)
                 .build();
     }
