+++ src/main/java/plate/back/global/initializer/ImageInitializer.java
@@ -1,25 +1,24 @@
 package plate.back.global.initializer;
-// package plate.back.initializer;
 
-// import org.springframework.boot.ApplicationArguments;
-// import org.springframework.boot.ApplicationRunner;
-// import org.springframework.stereotype.Component;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.ApplicationRunner;
+import org.springframework.stereotype.Component;
 
-// import lombok.RequiredArgsConstructor;
-// import plate.back.service.FileService;
+import lombok.RequiredArgsConstructor;
+import plate.back.global.s3.service.FileService;
 
-// @RequiredArgsConstructor
-// @Component
-// public class ImageInitializer implements ApplicationRunner {
+@RequiredArgsConstructor
+@Component
+public class ImageInitializer implements ApplicationRunner {
 
-// private final FileService fileService;
+    private final FileService fileService;
 
-// @Override
-// public void run(ApplicationArguments args) throws Exception {
-// // s3 파일 전체 삭제
-// fileService.deleteAll("total/vehicle");
-// fileService.deleteAll("total/plate");
-// fileService.deleteAll("relearn/vehicle");
-// fileService.deleteAll("relearn/plate");
-// }
-// }
+    @Override
+    public void run(ApplicationArguments args) throws Exception {
+        // s3 파일 전체 삭제
+        fileService.deleteAll("total/vehicle");
+        fileService.deleteAll("total/plate");
+        fileService.deleteAll("relearn/vehicle");
+        fileService.deleteAll("relearn/plate");
+    }
+}
